rstsr-core: relocate doc tests to doc_draft and gate entry binary
- Move doc-style tests from core_func/manuplication and linalg into a new
  tests/doc_draft/ module tree.
- Rename tests_core_row.rs to entry_row_cpu.rs and gate it on the
  row_major feature via [[test]] required-features in Cargo.toml, per
  ADR-0002 entry-binary test matrix; entry_row_faer / entry_col_cpu follow.
- Drop the standalone tests_issues.rs entry binary; test_issues is now
  included through entry_row_cpu.rs.
