+++ rstsr-core/tests/doc_draft/linalg/mod.rs
@@ -1,0 +1,3 @@
+pub mod test_vecdot;
+
+pub static CATEGORY: &str = "linalg";
