+++ rstsr-core/tests/doc_draft/manuplication/mod.rs
@@ -2,6 +2,8 @@ pub mod test_broadcast;
 pub mod test_broadcast_shapes;
 pub mod test_expand_dims;
 pub mod test_flip;
+pub mod test_into_compatible_shape;
+pub mod test_into_dim;
 pub mod test_moveaxis;
 pub mod test_reshape;
 pub mod test_squeeze;
