+++ rstsr-core/tests/doc_draft/mod.rs
@@ -1,0 +1,2 @@
+pub mod linalg;
+pub mod manuplication;
