--- rstsr-core/tests/tests_issues.rs
@@ -1,2 +1,0 @@
-/// Tests for issues reported in the rstsr repository.
-mod test_issues;
